backend/pyproject.toml
[project]
name = "ticketrush-backend"
version = "0.1.0"
description = "TicketRush FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "email-validator>=2.2.0",
    "apscheduler>=3.10.4",
    "qrcode[pil]>=7.4.2",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
